Some planning comments refined, initial movement for utils/files.lua and pickers.lua
