+++ lua/telekasten.lua
@@ -1,18 +1,14 @@
 local builtin = require("telescope.builtin")
 local actions = require("telescope.actions")
 local action_state = require("telescope.actions.state")
-local action_set = require("telescope.actions.set")
 local pickers = require("telescope.pickers")
 local finders = require("telescope.finders")
 local conf = require("telescope.config").values
-local scan = require("plenary.scandir")
 local utils = require("telescope.utils")
-local previewers = require("telescope.previewers")
 local make_entry = require("telescope.make_entry")
 local entry_display = require("telescope.pickers.entry_display")
 local sorters = require("telescope.sorters")
 local themes = require("telescope.themes")
-local debug_utils = require("plenary.debug_utils")
 local filetype = require("plenary.filetype")
 local taglinks = require("telekasten.utils.taglinks")
 local tagutils = require("telekasten.utils.tags")
@@ -30,48 +26,6 @@ local vim = vim
 
 local M = {}
 
--- string, string -> string
--- Returns the new note's file name, accounting for UUID usage and any desired space substition.
--- Move to utils/files.lua?
-local function generate_note_filename(uuid, title)
-    if config.options.filename_space_subst ~= nil then
-        title = title:gsub(" ", config.options.filename_space_subst)
-    end
-
-    local pp = Path:new(title)
-    local p_splits = pp:_split()
-    local filename = p_splits[#p_splits]
-    local subdir = title:gsub(tkutils.escape(filename), "")
-
-    local sep = config.options.uuid_sep or "-"
-    if config.options.new_note_filename ~= "uuid" and #title > 0 then
-        if config.options.new_note_filename == "uuid-title" then
-            return subdir .. uuid .. sep .. filename
-        elseif config.options.new_note_filename == "title-uuid" then
-            return title .. sep .. uuid
-        else
-            return title
-        end
-    else
-        return uuid
-    end
-end
-
--- string -> string
--- Returns the absolute path to a local file
--- Move to utils/files.lua
-local function make_config_path_absolute(path)
-    local ret = path
-    if not (Path:new(path):is_absolute()) and path ~= nil then
-        ret = config.options.home .. "/" .. path
-    end
-
-    if ret ~= nil then
-        ret = ret:gsub("/$", "")
-    end
-
-    return ret
-end
 
 -- string, string, string -> N/A
 -- No return, runs ripgrep and sed if and only if the global dir check passes
@@ -115,8 +69,8 @@ end
 
 -- N/A -> N/A
 -- No return
--- Saves all modified buffers if auto_set_filetype and buffer's filetype is telekasten
--- or do so if not auto_set_filetype
+-- Saves all modified buffers if auto_set_filetype and buffer's filetype is telekasten or if not auto_set_filetype
+-- Move to utils/files.lua? Arguably file related, but maybe better utils/init.lua
 local function save_all_mod_buffers()
     for i = 1, vim.fn.bufnr("$") do
         if
@@ -324,6 +278,7 @@ end
 
 -- string, string, string, string, table, function -> N/A
 -- No return, only creates a new note file from a given template file
+-- utils/templates.lua? utils/files.lua? utils/init.lua?
 local function create_note_from_template(
     title,
     uuid,
@@ -371,695 +326,6 @@ local function create_note_from_template(
     )
 end
 
---- Pinfo
---- Table of data related to a file's location and purpose
----    - fexists : true if file exists
----    - title : the title (filename including subdirs without extension)
----      - if opts.subdirs_in_links is false, no subdirs will be included
----    - filename : filename component only
----    - filepath : full path, identical to p
----    - root_dir : the root dir (home, dailies, ...)
----    - sub_dir : subdir if present, relative to root_dir
----    - is_daily_or_weekly : bool
----    - is_daily : bool
----    - is_weekly : bool
----    - template : suggested template based on opts
--- Move to utils/files.lua?
-local Pinfo = {
-    fexists = false,
-    title = "",
-    filename = "",
-    filepath = "",
-    root_dir = "",
-    sub_dir = "",
-    is_daily_or_weekly = false,
-    is_daily = false,
-    is_weekly = false,
-    template = "",
-    calendar_info = nil,
-}
-
--- table -> table
--- Returns a Pinfo table tailored to the file path or title specified in opts
--- Unsure of potential relation to new file paths and/or titles, check later
--- Keep with Pinfo
-function Pinfo:new(opts)
-    opts = opts or {}
-
-    local object = {}
-    setmetatable(object, self)
-    self.__index = self
-    if opts.filepath then
-        return object:resolve_path(opts.filepath, opts)
-    end
-    if opts.title ~= nil then
-        return object:resolve_link(opts.title, opts)
-    end
-    return object
-end
-
--- string, table -> table
---- resolve_path(p, opts)
---- inspects the path and returns a Pinfo table
--- Keep with Pinfo
-function Pinfo:resolve_path(p, opts)
-    opts = opts or {}
-    opts.subdirs_in_links = opts.subdirs_in_links
-        or config.options.subdirs_in_links
-
-    self.fexists = fileutils.file_exists(p)
-    self.filepath = p
-    self.root_dir = config.options.home
-    self.is_daily_or_weekly = false
-    self.is_daily = false
-    self.is_weekly = false
-
-    -- strip all dirs to get filename
-    local pp = Path:new(p)
-    local p_splits = pp:_split()
-    self.filename = p_splits[#p_splits]
-    self.title = self.filename:gsub(config.options.extension, "")
-
-    if vim.startswith(p, config.options.home) then
-        self.root_dir = config.options.home
-    end
-    if vim.startswith(p, config.options.dailies) then
-        self.root_dir = config.options.dailies
-        -- TODO: parse "title" into calendarinfo like in resolve_link
-        -- not really necessary as the file exists anyway and therefore we don't need to instantiate a template
-        self.is_daily_or_weekly = true
-        self.is_daily = true
-    end
-    if vim.startswith(p, config.options.weeklies) then
-        -- TODO: parse "title" into calendarinfo like in resolve_link
-        -- not really necessary as the file exists anyway and therefore we don't need to instantiate a template
-        self.root_dir = config.options.weeklies
-        self.is_daily_or_weekly = true
-        self.is_weekly = true
-    end
-
-    -- now work out subdir relative to root
-    self.sub_dir = p:gsub(tkutils.escape(self.root_dir .. "/"), "")
-        :gsub(tkutils.escape(self.filename), "")
-        :gsub("/$", "")
-        :gsub("^/", "")
-
-    if opts.subdirs_in_links and #self.sub_dir > 0 then
-        self.title = self.sub_dir .. "/" .. self.title
-    end
-
-    return self
-end
-
--- string -> bool, bool, table
--- Returns info on if the title given is for a daily or weekly and the date(s)
--- Maybe move to utils/dates.lua?
-local function check_if_daily_or_weekly(title)
-    local daily_match = "^(%d%d%d%d)-(%d%d)-(%d%d)$"
-    local weekly_match = "^(%d%d%d%d)-W(%d%d)$"
-
-    local is_daily = false
-    local is_weekly = false
-    local dateinfo = dateutils.calculate_dates(
-        nil,
-        config.options.calendar_opts.calendar_monday
-    ) -- sane default
-
-    -- Set return values for a daily note
-    local start, _, year, month, day = title:find(daily_match)
-    if start ~= nil then
-        if tonumber(month) < 13 then
-            if tonumber(day) < 32 then -- TODO: This should probably be refined for accuracy in 28-30 day months
-                is_daily = true
-                dateinfo.year = tonumber(year)
-                dateinfo.month = tonumber(month)
-                dateinfo.day = tonumber(day)
-                dateinfo = dateutils.calculate_dates(
-                    dateinfo,
-                    config.options.calendar_opts.calendar_monday
-                )
-            end
-        end
-    end
-
-    -- Set return values for a weekly note
-    -- Seems pointless to check both this and daily. Maybe try an else?
-    local week
-    start, _, year, week = title:find(weekly_match)
-    if start ~= nil then
-        if tonumber(week) < 53 then
-            is_weekly = true
-            -- ISO8601 week -> date calculation
-            dateinfo = dateutils.isoweek_to_date(tonumber(year), tonumber(week))
-            dateinfo = dateutils.calculate_dates(
-                dateinfo,
-                config.options.calendar_opts.calendar_monday
-            )
-        end
-    end
-
-    return is_daily, is_weekly, dateinfo
-end
-
--- string, table -> table
--- Returns a Pinfo table for the given title and options
-function Pinfo:resolve_link(title, opts)
-    -- Set options, preferring passed opts over user config
-    opts = opts or {}
-    opts.weeklies = opts.weeklies or config.options.weeklies
-    opts.dailies = opts.dailies or config.options.dailies
-    opts.home = opts.home or config.options.home
-    opts.extension = opts.extension or config.options.extension
-    opts.template_handling = opts.template_handling
-        or config.options.template_handling
-    opts.new_note_location = opts.new_note_location
-        or config.options.new_note_location
-
-    -- Set basic Pinfo values
-    self.fexists = false
-    self.title = title
-    self.filename = title .. opts.extension
-    self.filename = self.filename:gsub("^%./", "") -- strip potential leading ./
-    self.root_dir = opts.home
-    self.is_daily_or_weekly = false
-    self.is_daily = false
-    self.is_weekly = false
-    self.template = nil
-    self.calendar_info = nil
-
-    -- Try checking for existence and assigning values as a weekly, then as a daily, then as a plain note in home
-    if
-        opts.weeklies
-        and fileutils.file_exists(opts.weeklies .. "/" .. self.filename)
-    then
-        -- TODO: parse "title" into calendarinfo like below
-        -- not really necessary as the file exists anyway and therefore we don't need to instantiate a template
-        -- if we still want calendar_info, just move the code for it out of `if self.fexists == false`.
-        self.filepath = opts.weeklies .. "/" .. self.filename
-        self.fexists = true
-        self.root_dir = opts.weeklies
-        self.is_daily_or_weekly = true
-        self.is_weekly = true
-    end
-    if -- TODO: This should be able to convert to an elseif, I think. Weekly and daily file names are distinct
-        opts.dailies
-        and fileutils.file_exists(opts.dailies .. "/" .. self.filename)
-    then
-        -- TODO: parse "title" into calendarinfo like below
-        -- not really necessary as the file exists anyway and therefore we don't need to instantiate a template
-        -- if we still want calendar_info, just move the code for it out of `if self.fexists == false`.
-        self.filepath = opts.dailies .. "/" .. self.filename
-        self.fexists = true
-        self.root_dir = opts.dailies
-        self.is_daily_or_weekly = true
-        self.is_daily = true
-    end
-    if fileutils.file_exists(opts.home .. "/" .. self.filename) then
-        self.filepath = opts.home .. "/" .. self.filename
-        self.fexists = true
-    end
-
-    -- If still nothing found, check subdirectories
-    if self.fexists == false then
-        -- now search for it in all subdirs
-        local subdirs = scan.scan_dir(opts.home, { only_dirs = true })
-        local tempfn
-        for _, folder in pairs(subdirs) do
-            tempfn = folder .. "/" .. self.filename
-            -- [[testnote]]
-            if fileutils.file_exists(tempfn) then
-                self.filepath = tempfn
-                self.fexists = true
-                -- print("Found: " ..self.filename)
-                break
-            end
-        end
-    end
-
-    -- if we just cannot find the note, check if it's a daily or weekly one
-    -- Note that fexists is not changed; This is prep in case a new note is made
-    if self.fexists == false then
-        -- TODO: if we're not smart, we also shouldn't need to try to set the calendar info..?
-        --       I bet someone will want the info in there, so let's put it in if possible
-        _, _, self.calendar_info = check_if_daily_or_weekly(self.title) -- will set today as default, so leave in!
-
-        if opts.new_note_location == "smart" then
-            self.filepath = opts.home .. "/" .. self.filename -- default
-            self.is_daily, self.is_weekly, self.calendar_info =
-                check_if_daily_or_weekly(self.title) -- TODO: Don't replicate call, simply save all values above
-            if self.is_daily == true then
-                self.root_dir = opts.dailies
-                self.filepath = opts.dailies .. "/" .. self.filename
-                self.is_daily_or_weekly = true
-            end
-            if self.is_weekly == true then
-                self.root_dir = opts.weeklies
-                self.filepath = opts.weeklies .. "/" .. self.filename
-                self.is_daily_or_weekly = true
-            end
-        elseif opts.new_note_location == "same_as_current" then
-            local cwd = vim.fn.expand("%:p")
-            if #cwd > 0 then
-                self.root_dir = Path:new(cwd):parent():absolute()
-                if Path:new(self.root_dir):exists() then
-                    -- check if potential subfolders in filename would end up in a non-existing directory
-                    self.filepath = self.root_dir .. "/" .. self.filename
-                    if not Path:new(self.filepath):parent():exists() then
-                        print("Path " .. self.filepath .. " is invalid!")
-                        -- self.filepath = opts.home .. "/" .. self.filename
-                    end
-                else
-                    print("Path " .. self.root_dir .. " is invalid!")
-                    -- self.filepath = opts.home .. "/" .. self.filename
-                end
-            else
-                self.filepath = opts.home .. "/" .. self.filename
-            end
-        else
-            -- default fn for creation
-            self.filepath = opts.home .. "/" .. self.filename
-        end
-
-        -- final round, there still can be a subdir mess-up
-        if not Path:new(self.filepath):parent():exists() then
-            print("Path " .. self.filepath .. " is invalid!")
-        end
-    end
-
-    -- now work out subdir relative to root
-    self.sub_dir = self.filepath
-        :gsub(tkutils.escape(self.root_dir .. "/"), "")
-        :gsub(tkutils.escape(self.filename), "")
-        :gsub("/$", "")
-        :gsub("^/", "")
-
-    -- now suggest a template based on opts
-    self.template = M.note_type_templates.normal
-    if opts.template_handling == "prefer_new_note" then
-        self.template = M.note_type_templates.normal
-    elseif opts.template_handling == "always_ask" then
-        self.template = nil
-    elseif opts.template_handling == "smart" then
-        if self.is_daily then
-            self.template = M.note_type_templates.daily
-        elseif self.is_weekly then
-            self.template = M.note_type_templates.weekly
-        else
-            self.template = M.note_type_templates.normal
-        end
-    end
-
-    return self
-end
-
--- table, table -> table
--- Returns all entries in flist with filetypes in ftypes
-local function filter_filetypes(flist, ftypes)
-    local new_fl = {}
-    ftypes = ftypes or { config.options.extension }
-
-    local ftypeok = {}
-    for _, ft in pairs(ftypes) do
-        ftypeok[ft] = true
-    end
-
-    for _, fn in pairs(flist) do
-        if ftypeok[fileutils.get_extension(fn)] then
-            table.insert(new_fl, fn)
-        end
-    end
-    return new_fl
-end
-
--- Defines how to  provide a media preview when trying to find files
--- TODO: This is all setup for defining media_preview just for use in find_files_sorted, feels like it could be cleaner
-local sourced_file = debug_utils.sourced_filepath()
-M.base_directory = vim.fn.fnamemodify(sourced_file, ":h:h:h")
-local media_files_base_directory = M.base_directory
-    .. "/telescope-media-files.nvim"
-local defaulter = utils.make_default_callable
-local media_preview = defaulter(function(opts)
-    local preview_cmd = ""
-    if config.options.media_previewer == "telescope-media-files" then
-        preview_cmd = media_files_base_directory .. "/scripts/vimg"
-    end
-
-    if config.options.media_previewer == "catimg-previewer" then
-        preview_cmd = M.base_directory
-            .. "/telekasten.nvim/scripts/catimg-previewer"
-    end
-
-    if vim.startswith(config.options.media_previewer, "viu-previewer") then
-        preview_cmd = M.base_directory
-            .. "/telekasten.nvim/scripts/"
-            .. config.options.media_previewer
-    end
-
-    if vim.fn.executable(preview_cmd) == 0 then
-        print("Previewer not found: " .. preview_cmd)
-        return conf.file_previewer(opts)
-    end
-    return previewers.new_termopen_previewer({
-        get_command = opts.get_command
-            or function(entry)
-                local tmp_table = vim.split(entry.value, "\t")
-                local preview = opts.get_preview_window()
-                opts.cwd = opts.cwd and vim.fn.expand(opts.cwd)
-                    or vim.loop.cwd()
-                if vim.tbl_isempty(tmp_table) then
-                    return { "echo", "" }
-                end
-                print(tmp_table[1])
-                return {
-                    preview_cmd,
-                    tmp_table[1],
-                    preview.col,
-                    preview.line + 1,
-                    preview.width,
-                    preview.height,
-                }
-            end,
-    })
-end, {})
-
--- note picker actions
--- Move to utils/files.lua? Make sure to bring table entry definitions from below
-local picker_actions = {}
-
--- find_files_sorted(opts)
--- like builtin.find_files, but:
---     - uses plenary.scan_dir synchronously instead of external jobs
---     - pre-sorts the file list in descending order (nice for dates, most recent first)
---     - filters for allowed file types by extension
---     - (also supports devicons)
---     - displays subdirs if necessary
---         - e.g. when searching for daily notes, no subdirs are displayed
---         - but when entering a date in find_notes, the daily/ and weekly/ subdirs are displayed
---     - optionally previews media (pdf, images, mp4, webm)
---         - this requires the telescope-media-files.nvim extension
--- Move to utils/files.lua
-local function find_files_sorted(opts)
-    opts = opts or {}
-    local search_pattern = opts.search_pattern or nil
-    local search_depth = opts.search_depth or nil
-    local scan_opts = { search_pattern = search_pattern, depth = search_depth }
-
-    local file_list = scan.scan_dir(opts.cwd, scan_opts)
-    local filter_extensions = opts.filter_extensions
-        or config.options.filter_extensions
-    file_list = filter_filetypes(file_list, filter_extensions)
-    local sort_option = opts.sort or "filename"
-    if sort_option == "modified" then
-        table.sort(file_list, function(a, b)
-            return vim.fn.getftime(a) > vim.fn.getftime(b)
-        end)
-    else
-        table.sort(file_list, function(a, b)
-            return a > b
-        end)
-    end
-
-    local counts = nil
-    if opts.show_link_counts then
-        counts = linkutils.generate_backlink_map(config.options)
-    end
-
-    -- display with devicons
-    local function iconic_display(display_entry)
-        local display_opts = {
-            path_display = function(_, e)
-                return e:gsub(tkutils.escape(opts.cwd .. "/"), "")
-            end,
-        }
-
-        local hl_group
-        local display = utils.transform_path(display_opts, display_entry.value)
-
-        display, hl_group =
-            utils.transform_devicons(display_entry.value, display, false)
-
-        if hl_group then
-            return display, { { { 1, 3 }, hl_group } }
-        else
-            return display
-        end
-    end
-
-    -- for media_files
-    local popup_opts = {}
-    opts.get_preview_window = function()
-        return popup_opts.preview
-    end
-
-    -- local width = config.width
-    --     or config.layout_config.width
-    --     or config.layout_config[config.layout_strategy].width
-    --     or vim.o.columns
-    -- local telescope_win_width
-    -- if width > 1 then
-    --     telescope_win_width = width
-    -- else
-    --     telescope_win_width = math.floor(vim.o.columns * width)
-    -- end
-    local displayer = entry_display.create({
-        separator = "",
-        items = {
-            { width = 4 },
-            { width = 4 },
-            { remaining = true },
-        },
-    })
-
-    local function make_display(entry)
-        local fn = entry.value
-        local nlinks = counts.link_counts[fn] or 0
-        local nbacks = counts.backlink_counts[fn] or 0
-
-        if opts.show_link_counts then
-            local display, hl = iconic_display(entry)
-
-            return displayer({
-                {
-                    "L" .. tostring(nlinks),
-                    function()
-                        return {
-                            { { 0, 1 }, "tkTagSep" },
-                            { { 1, 3 }, "tkTag" },
-                        }
-                    end,
-                },
-                {
-                    "B" .. tostring(nbacks),
-                    function()
-                        return {
-                            { { 0, 1 }, "tkTagSep" },
-                            { { 1, 3 }, "DevIconMd" },
-                        }
-                    end,
-                },
-                {
-                    display,
-                    function()
-                        return hl
-                    end,
-                },
-            })
-        else
-            return iconic_display(entry)
-        end
-    end
-
-    local function entry_maker(entry)
-        local iconic_entry = {}
-        iconic_entry.value = entry
-        iconic_entry.path = entry
-        iconic_entry.ordinal = entry
-        if opts.show_link_counts then
-            iconic_entry.display = make_display
-        else
-            iconic_entry.display = iconic_display
-        end
-        return iconic_entry
-    end
-
-    local previewer = conf.file_previewer(opts)
-    if opts.preview_type == "media" then
-        previewer = media_preview.new(opts)
-    end
-
-    opts.attach_mappings = opts.attach_mappings
-        or function(_, _)
-            actions.select_default:replace(picker_actions.select_default)
-        end
-
-    local picker = pickers.new(opts, {
-        finder = finders.new_table({
-            results = file_list,
-            entry_maker = entry_maker,
-        }),
-        sorter = conf.generic_sorter(opts),
-        previewer = previewer,
-    })
-
-    -- local oc = picker.finder.close
-    --
-    -- picker.finder.close = function()
-    --     print('on close')
-    --     print(vim.inspect(picker:get_selection()))
-    --     -- unfortunately, no way to tell if the selection was confirmed or
-    --     -- canceled out
-    --     oc()
-    --     -- alternative: attach default mappings for <ESC> and <C-c>
-    --     --       if anyone quits with q!, it's their fault
-    -- end
-
-    -- for media_files:
-    local line_count = vim.o.lines - vim.o.cmdheight
-    if vim.o.laststatus ~= 0 then
-        line_count = line_count - 1
-    end
-
-    popup_opts = picker:get_window_options(vim.o.columns, line_count)
-
-    picker:find()
-end
-
--- N/A -> N/A
--- No return. If user config auto sets filetype or syntax, actually set them
--- Keep with picker_actions definition
-picker_actions.post_open = function()
-    if config.options.auto_set_filetype then
-        vim.cmd("set ft=telekasten")
-    end
-    if config.options.auto_set_syntax then
-        vim.cmd("set syntax=telekasten")
-    end
-end
-
--- int -> nil? action_set.select returns result of action_set.edit, but edit doesn't return a value
--- Keep with picker_actions definition
-picker_actions.select_default = function(prompt_bufnr)
-    local ret = action_set.select(prompt_bufnr, "default")
-    picker_actions.post_open()
-    return ret
-end
-
--- table -> function
--- Returns a configured function mapped to a key while using the picker
-function picker_actions.close(opts)
-    opts = opts or {}
-    return function(prompt_bufnr)
-        actions.close(prompt_bufnr)
-        if opts.erase then
-            if fileutils.file_exists(opts.erase_file) then
-                vim.fn.delete(opts.erase_file)
-            end
-        end
-    end
-end
-
--- table -> function
--- Returns a configured function that gets mapped to a key while using the picker in FindAllTags
-function picker_actions.paste_tag(opts)
-    return function(prompt_bufnr)
-        actions.close(prompt_bufnr)
-        local selection = action_state.get_selected_entry()
-        vim.api.nvim_put({ selection.value.tag }, "", true, true)
-        if opts.insert_after_inserting or opts.i then
-            vim.api.nvim_feedkeys("A", "m", false)
-        end
-    end
-end
-
--- table -> function
--- Returns a configured function that gets mapped to a key while using the picker in FindAllTags
-function picker_actions.yank_tag(opts)
-    return function(prompt_bufnr)
-        opts = opts or {}
-        if opts.close_after_yanking then
-            actions.close(prompt_bufnr)
-        end
-        local selection = action_state.get_selected_entry()
-        vim.fn.setreg('"', selection.value.tag)
-        print("yanked " .. selection.value.tag)
-    end
-end
-
--- table -> function
--- Returns a configured function that gets mapped to a key while using the picker
-function picker_actions.paste_link(opts)
-    opts = opts or {}
-    opts.subdirs_in_links = opts.subdirs_in_links
-        or config.options.subdirs_in_links
-    return function(prompt_bufnr)
-        actions.close(prompt_bufnr)
-        local selection = action_state.get_selected_entry()
-        local pinfo = Pinfo:new({
-            filepath = selection.filename or selection.value,
-            opts,
-        })
-        local title = "[[" .. pinfo.title .. "]]"
-        vim.api.nvim_put({ title }, "", true, true)
-        if opts.insert_after_inserting or opts.i then
-            vim.api.nvim_feedkeys("A", "m", false)
-        end
-    end
-end
-
--- table -> function
--- Returns a configured function that gets mapped to a key while using the picker
-function picker_actions.yank_link(opts)
-    return function(prompt_bufnr)
-        opts = opts or {}
-        opts.subdirs_in_links = opts.subdirs_in_links
-            or config.options.subdirs_in_links
-        if opts.close_after_yanking then
-            actions.close(prompt_bufnr)
-        end
-        local selection = action_state.get_selected_entry()
-        local pinfo = Pinfo:new({
-            filepath = selection.filename or selection.value,
-            opts,
-        })
-        local title = "[[" .. pinfo.title .. "]]"
-        vim.fn.setreg('"', title)
-        print("yanked " .. title)
-    end
-end
-
--- table -> function
--- Returns a configured function that gets mapped to a key while using the picker for images
-function picker_actions.paste_img_link(opts)
-    return function(prompt_bufnr)
-        actions.close(prompt_bufnr)
-        local selection = action_state.get_selected_entry()
-        local fn = selection.value
-        fn = make_relative_path(vim.fn.expand("%:p"), fn, "/")
-        local imglink = "![](" .. fn .. ")"
-        vim.api.nvim_put({ imglink }, "", true, true)
-        if opts.insert_after_inserting or opts.i then
-            vim.api.nvim_feedkeys("A", "m", false)
-        end
-    end
-end
-
--- table -> function
--- Returns a configured function that gets mapped to a key while using the picker for images
-function picker_actions.yank_img_link(opts)
-    return function(prompt_bufnr)
-        opts = opts or {}
-        if opts.close_after_yanking then
-            actions.close(prompt_bufnr)
-        end
-        local selection = action_state.get_selected_entry()
-        local fn = selection.value
-        fn = make_relative_path(vim.fn.expand("%:p"), fn, "/")
-        local imglink = "![](" .. fn .. ")"
-        vim.fn.setreg('"', imglink)
-        print("yanked " .. imglink)
-    end
-end
-
 --
 -- FindDailyNotes:
 -- ---------------
@@ -1068,7 +334,7 @@ end
 --
 -- table -> N/A
 -- No return. Opens a picker looking for daily notes, creating new from template if needed
--- Move to utils/files.lua? Arguably file related
+-- USER FACING, leave in place
 local function FindDailyNotes(opts)
     opts = opts or {}
     opts.insert_after_inserting = opts.insert_after_inserting
@@ -1145,7 +411,7 @@ end
 --
 -- table -> N/A
 -- No return. Defines a picker looking for weekly notes, creating a new one from template if needed
--- Move to utils/files.lua? Arguably file related. Keep with FindDailyNotes either way.
+-- USER FACING, leave in place
 local function FindWeeklyNotes(opts)
     opts = opts or {}
     opts.insert_after_inserting = opts.insert_after_inserting
@@ -1223,6 +489,7 @@ end
 -- Select from all notes and put a link in the current buffer
 -- table -> N/A
 -- No return. Sets up a picker from which users can pick a note to insert a link to
+-- USER FACING, leave in place
 local function InsertLink(opts)
     opts = opts or {}
     opts.insert_after_inserting = opts.insert_after_inserting
@@ -1505,7 +772,7 @@ end
 
 -- string, string -> N/A
 -- No return. Update links with name change if configured to
--- Move to utils/files.lua
+-- Move to utils/files.lua? utils/links.lua?
 local function rename_update_links(oldfile, newname)
     if config.options.rename_update_links == true then
         -- Only look for the first part of the link, so we do not touch to #heading or #^paragraph
@@ -1614,7 +881,7 @@ end
 -- find note for date and create it if necessary.
 --
 -- table -> N/A
--- Move to utils/files.lua?
+-- Move to utils/files.lua? Technically not user facing...
 local function GotoDate(opts)
     opts.dates = dateutils.calculate_dates(
         opts.date_table,
@@ -1918,7 +1185,7 @@ end
 -- Prompts for title, then pops up telescope for template selection,
 -- creates the new note by template and opens it
 -- table, string -> N/A
--- Move? Used by CreateNoteSelectTemplate and FollowLink
+-- Move? Used by CreateNoteSelectTemplate and FollowLink. Maybe good case for utils/init.lua to share it?
 local function on_create_with_template(opts, title)
     if title == nil then
         return
@@ -2601,7 +1868,7 @@ local function GotoThisWeek(opts)
                 end
                 vim.cmd("e " .. fname)
             else
-                find_files_sorted({
+                fileutils.find_files_sorted({
                     prompt_title = "Goto this week:",
                     cwd = config.options.weeklies,
                     default_text = title,
@@ -3167,21 +2434,6 @@ TelekastenCmd.command = function(subcommand)
             })
         end
         show(theme)
-    end
-end
-
--- table -> function
--- Keep with other picker_actions
-function picker_actions.create_new(opts)
-    opts = opts or {}
-    opts.subdirs_in_links = opts.subdirs_in_links
-        or config.options.subdirs_in_links
-    return function(prompt_bufnr)
-        local prompt =
-            action_state.get_current_picker(prompt_bufnr).sorter._discard_state.prompt
-        actions.close(prompt_bufnr)
-        on_create(opts, prompt)
-        -- local selection = action_state.get_selected_entry()
     end
 end
